# gcs: Support independent fp rate and bin size.